Updated with my changes so far
The main file that needs to be edited is “baps.php”. This file contains the hard-coded info for which companies and time slots are in the system. I added a variable $bool_initialize_db to set whether or not (True/False) the plugin will add new companies/timeslots/fields to the database when the plugin is activated. So set this to True after you have cleared the database tables, but set it to False if you want to make changes afterwards to the UI, etc so that you don’t generate duplicant companies/fields, etc.
Once the list of companies and the timeslots were known for the event, some code regions in “baps.php” were updated. Around line 163, the list after  “$query = "INSERT IGNORE INTO {$wp}baps_companies” was updated as well the time slots below. Then, there are if/elseif statements for all of the companies with their hardcoded time slots that also need to be updated based on whether the companies have silver/gold packages.
I also modified “baps-ui.php” to replace the feedback link in the registration messages with a new google form for this year (https://forms.gle/oroTqTiw89txgsZy6 ) . I also replaced the “vienna@best-eu.org” with “orion.forowycz@best-eu.org” (my email) so that the whole LBG doesn’t get spammed if someone were to reply to one of the emails that gets sent out.